changed default value for Configuration Directory
Initial the default value for the configuration directory is set to the path where the module is getting installed. However it would be better to have the directory outside the modules Folder. This change will expect the configurations folder on the same directory level where the sources folder is located.